requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
google-genai
requests
pypdf
aiohttp
aioredis
asyncio-throttle
aiofiles
python-docx
transformers
torch
edge-tts
numpy
scikit-learn
